refactor cli and runtime

+++ cli/create/create.ts
@@ -1,21 +1,30 @@
-import { stdFs, stdPath } from "../../deps.ts";
-import { CLICommand, CLICommandOptionDataType, optionalFileOption } from "../cli.ts";
+import { Command, stdFs } from "../../deps.ts";
 import type { CLIContext } from "../context.ts";
 
-async function run(context?: CLIContext): Promise<void> {
-    const filePath = stdPath.join('.', context?.args.file ?? stdPath.join('build', 'build.ts'));
-    const override = context?.args.override ?? false;
+export function getCreateCommand(): { cmd: Command, buildContextRequired: boolean, action: (context: CLIContext) => Promise<void> } {
+    return {
+        cmd: new Command()
+            .description('Create a new build file')
+            .option('--override', 'Override an existing build file.', { default: false }),
+        buildContextRequired: false,
+        action: async (context: CLIContext) => {
+            if (context.buildFile === undefined) {
+                throw new Error('Build file is unavailable.');
+            }
 
-    if (await stdFs.exists(filePath)) {
-        if (!override) {
-            throw new Error(`Build file already exists at '${filePath}'.`);
-        }
+            const filePath = context.buildFile;
+            const override = context.args['override'] ?? false;
 
-        await Deno.remove(filePath);
-    }
+            if (await stdFs.exists(filePath)) {
+                if (!override) {
+                    throw new Error(`Build file already exists at '${filePath}'.`);
+                }
 
-    await stdFs.ensureFile(filePath);
-    await Deno.writeFile(filePath, new TextEncoder().encode(
+                await Deno.remove(filePath);
+            }
+
+            await stdFs.ensureFile(filePath);
+            await Deno.writeFile(filePath, new TextEncoder().encode(
 `import { createBuilder, task } from "https://deno.land/x/denogent/lib/mod.ts";
 
 const someTask = task('some task')
@@ -31,23 +40,7 @@ createBuilder({
 });
 `));
 
-    context?.logger('info', `Created '${filePath}'.`);
-}
-
-export function createCommandDescription(): CLICommand {
-    return {
-        name: 'create',
-        description: 'Create a new build file',
-        options: [
-            optionalFileOption,
-            {
-                name: 'override',
-                description: 'Override an existing build file',
-                dataType: CLICommandOptionDataType.Boolean,
-                required: false
-            }
-        ],
-        requireBuildContext: false,
-        fn: run
+            context.runtime.loggerFn('info', `Created '${filePath}'.`);
+        }
     };
 }
